Allow recursion in brainstorm example (#850)
Demo recursive schema support in Brainstorm. As far as I know this is the first non-trivial use of the new recursive schema APIs, so this was a valuable test of those, as well as a demo for them.

Additionally, more functions were moved onto the schema as methods, root and group handling were unified (via Items), some comments were converted to doc comments to help with IntelliSense and show off that doc comments on fields in schema actually work.

A few places stopped taking in parents explicitly and instead look them up in the tree.

Tree.is has been used instead of instancof since it works in more cases (for example with leaf nodes and POJO mode schema): this is more of just encouraging a more general pattern than a fix since instancof works fine where it was used.